+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         <activity android:name=".admin.AdminActivity"></activity>
         <activity android:name=".analysis.team.TeamAnalysisActivity"></activity>
         <activity android:name=".analysis.attribute.AttributeAnalysisActivity"></activity>
-        <activity android:name=".DuckActivity"></activity>
+        <activity android:name=".duck.DuckActivity"></activity>
 
         <provider
             android:name="androidx.core.content.FileProvider"
