+++ app/src/main/java/com/frc107/scouting2019/admin/AdminViewModel.java
@@ -3,7 +3,7 @@ package com.frc107.scouting2019.admin;
 import android.content.Context;
 import android.net.Uri;
 
-import com.frc107.scouting2019.IUIListener;
+import com.frc107.scouting2019.ui.IUIListener;
 
 import java.io.File;
 import java.util.ArrayList;
