+++ app/src/main/java/com/frc107/scouting2019/analysis/attribute/AttributeAnalysisActivity.java
@@ -4,8 +4,8 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
 
-import com.frc107.scouting2019.BaseActivity;
-import com.frc107.scouting2019.IUIListener;
+import com.frc107.scouting2019.ui.BaseActivity;
+import com.frc107.scouting2019.ui.IUIListener;
 
 import android.view.View;
 import android.widget.ListView;
