+++ app/src/main/java/com/frc107/scouting2019/analysis/team/TeamAnalysisModel.java
@@ -3,7 +3,7 @@ package com.frc107.scouting2019.analysis.team;
 import android.util.SparseArray;
 
 import com.frc107.scouting2019.analysis.IAnalysisListener;
-import com.frc107.scouting2019.IUIListener;
+import com.frc107.scouting2019.ui.IUIListener;
 import com.frc107.scouting2019.analysis.LoadDataTask;
 import com.frc107.scouting2019.analysis.TeamDetails;
