+++ app/src/main/java/com/frc107/scouting2019/cycle/CycleModel.java
@@ -2,12 +2,12 @@ package com.frc107.scouting2019.cycle;
 
 import com.frc107.scouting2019.R;
 import com.frc107.scouting2019.Scouting;
-import com.frc107.scouting2019.ScoutModel;
-import com.frc107.scouting2019.question.Question;
-import com.frc107.scouting2019.question.RadioQuestion;
-import com.frc107.scouting2019.question.ToggleQuestion;
+import com.frc107.scouting2019.model.BaseModel;
+import com.frc107.scouting2019.model.question.Question;
+import com.frc107.scouting2019.model.question.RadioQuestion;
+import com.frc107.scouting2019.model.question.ToggleQuestion;
 
-public class CycleModel extends ScoutModel {
+public class CycleModel extends BaseModel {
     private int cycleNum;
     private boolean isFirstCycle = true;
     private boolean hasUsedStartingItem;
