+++ app/src/main/java/com/frc107/scouting2019/cycle/CycleViewModel.java
@@ -1,8 +1,8 @@
 package com.frc107.scouting2019.cycle;
 
-import com.frc107.scouting2019.ScoutViewModel;
+import com.frc107.scouting2019.viewmodel.BaseViewModel;
 
-public class CycleViewModel extends ScoutViewModel {
+public class CycleViewModel extends BaseViewModel {
     public CycleViewModel(int teamNumber) {
         model = new CycleModel(teamNumber);
     }
