+++ app/src/main/java/com/frc107/scouting2019/duck/DuckActivity.java
@@ -1,8 +1,8 @@
-package com.frc107.scouting2019;
+package com.frc107.scouting2019.duck;
 
 import android.os.Bundle;
 
-import com.frc107.scouting2019.BaseActivity;
+import com.frc107.scouting2019.ui.BaseActivity;
 import com.frc107.scouting2019.R;
 
 public class DuckActivity extends BaseActivity {
