+++ app/src/main/java/com/frc107/scouting2019/model/question/NumberQuestion.java
@@ -1,4 +1,4 @@
-package com.frc107.scouting2019.question;
+package com.frc107.scouting2019.model.question;
 
 import java.util.ArrayList;
