+++ app/src/main/java/com/frc107/scouting2019/pit/PitModel.java
@@ -5,15 +5,15 @@ import android.util.Log;
 
 import com.frc107.scouting2019.R;
 import com.frc107.scouting2019.Scouting;
-import com.frc107.scouting2019.ScoutModel;
-import com.frc107.scouting2019.question.Question;
-import com.frc107.scouting2019.question.RadioQuestion;
-import com.frc107.scouting2019.question.TextQuestion;
+import com.frc107.scouting2019.model.BaseModel;
+import com.frc107.scouting2019.model.question.Question;
+import com.frc107.scouting2019.model.question.RadioQuestion;
+import com.frc107.scouting2019.model.question.TextQuestion;
 
 import java.io.File;
 import java.io.IOException;
 
-public class PitModel extends ScoutModel {
+public class PitModel extends BaseModel {
     public PitModel() {
         super();
         setFileNameHeader("Pit");
