+++ app/src/main/java/com/frc107/scouting2019/pit/PitViewModel.java
@@ -1,10 +1,10 @@
 package com.frc107.scouting2019.pit;
 
-import com.frc107.scouting2019.ScoutViewModel;
+import com.frc107.scouting2019.viewmodel.BaseViewModel;
 
 import java.io.File;
 
-public class PitViewModel extends ScoutViewModel {
+public class PitViewModel extends BaseViewModel {
     public PitViewModel() {
         model = new PitModel();
     }
