+++ app/src/main/java/com/frc107/scouting2019/ui/IUIListener.java
@@ -1,4 +1,4 @@
-package com.frc107.scouting2019;
+package com.frc107.scouting2019.ui;
 
 public interface IUIListener {
     void callback(boolean error);
